requirement.txt
streamlit
python-dotenv # for working on load_dotenv
google-genai
google-generativeai
pillow
gTTS